+++ pkg/policy/rule_test.go
@@ -1789,7 +1789,7 @@ func (ds *PolicyTestSuite) TestIngressL4AllowAll(c *C) {
 
 	filter, ok := l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 1)
@@ -1834,8 +1834,9 @@ func (ds *PolicyTestSuite) TestIngressL4AllowAllNamedPort(c *C) {
 
 	filter, ok := l4IngressPolicy["port-80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 0)
+	c.Assert(filter.Port, Equals, uint16(0))
 	c.Assert(filter.PortName, Equals, "port-80")
+
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 1)
@@ -1906,7 +1907,7 @@ func (ds *PolicyTestSuite) TestEgressL4AllowAll(c *C) {
 
 	filter, ok := l4EgressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, false)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 1)
@@ -1962,7 +1963,7 @@ func (ds *PolicyTestSuite) TestEgressL4AllowWorld(c *C) {
 
 	filter, ok := l4EgressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, false)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 3)
@@ -2017,7 +2018,7 @@ func (ds *PolicyTestSuite) TestEgressL4AllowAllEntity(c *C) {
 
 	filter, ok := l4EgressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, false)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 1)
@@ -2192,7 +2193,7 @@ func (ds *PolicyTestSuite) TestL4WildcardMerge(c *C) {
 
 	filter, ok := l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 2)
@@ -2219,7 +2220,7 @@ func (ds *PolicyTestSuite) TestL4WildcardMerge(c *C) {
 
 	filterL7, ok := l4IngressPolicy["7000/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filterL7.Port, Equals, 7000)
+	c.Assert(filterL7.Port, Equals, uint16(7000))
 	c.Assert(filterL7.Ingress, Equals, true)
 
 	c.Assert(len(filterL7.PerSelectorPolicies), Equals, 1)
@@ -2299,7 +2300,7 @@ func (ds *PolicyTestSuite) TestL4WildcardMerge(c *C) {
 
 	filter, ok = l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 2)
@@ -2310,7 +2311,7 @@ func (ds *PolicyTestSuite) TestL4WildcardMerge(c *C) {
 
 	filterL7, ok = l4IngressPolicy["7000/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filterL7.Port, Equals, 7000)
+	c.Assert(filterL7.Port, Equals, uint16(7000))
 	c.Assert(filterL7.Ingress, Equals, true)
 
 	c.Assert(len(filterL7.PerSelectorPolicies), Equals, 1)
@@ -2364,7 +2365,7 @@ func (ds *PolicyTestSuite) TestL4WildcardMerge(c *C) {
 
 	filter, ok = l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(filter.L7Parser, Equals, ParserTypeHTTP)
@@ -2418,7 +2419,7 @@ func (ds *PolicyTestSuite) TestL4WildcardMerge(c *C) {
 
 	filter, ok = l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(filter.L7Parser, Equals, ParserTypeHTTP)
@@ -2477,7 +2478,7 @@ func (ds *PolicyTestSuite) TestL3L4L7Merge(c *C) {
 
 	filter, ok := l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(len(filter.PerSelectorPolicies), Equals, 2)
@@ -2546,7 +2547,7 @@ func (ds *PolicyTestSuite) TestL3L4L7Merge(c *C) {
 
 	filter, ok = l4IngressPolicy["80/TCP"]
 	c.Assert(ok, Equals, true)
-	c.Assert(filter.Port, Equals, 80)
+	c.Assert(filter.Port, Equals, uint16(80))
 	c.Assert(filter.Ingress, Equals, true)
 
 	c.Assert(filter.L7Parser, Equals, ParserTypeHTTP)
